pyproject.toml
[project]
name = "grit-lang-chain-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "chromadb>=1.5.4",
    "docx2txt>=0.9",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.17",
    "langchain-ollama>=1.0.1",
    "langchain-text-splitters>=1.0.0",
    "openpyxl>=3.1.5",
    "pypdf>=6.7.5",
    "python-pptx>=1.0.2",
    "tiktoken>=0.12.0",
]
